post

+++ app/Http/Controllers/TasksController.php
@@ -16,15 +16,29 @@ class TasksController extends Controller
    
     public function index()
     {
-         // メッセージ一覧を取得
-        $tasks = Task::all();
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            
+            $user = \Auth::user();
+            
+            
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->task()->orderBy('created_at', 'desc')->paginate(10);
+            
+             // タスク一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+                
+            ]);
+            
+        }else{    
+              
+            return redirect('login');
+    
+        }
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -38,6 +52,8 @@ class TasksController extends Controller
         // メッセージ作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
+
+            
         ]);
     }
 
@@ -57,13 +73,15 @@ class TasksController extends Controller
             'status' => 'required|max:10',  
         ]);
         
+     
         $task = new Task;
-        $task->content = $request->content;
+        $task->user_id = $request->user()->id;
         $task->status = $request->status;
+        $task->content = $request->content;
         $task->save();
-
-        // トップページへリダイレクトさせる
+        
         return redirect('/');
+       
     }
 
     /**
@@ -77,7 +95,7 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        // タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -136,11 +154,16 @@ class TasksController extends Controller
    
     public function destroy($id)
     {
-         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        
         return redirect('/');
     }
 }
